refactor: migrate Exa search from toolkit to standalone functions
