+++ packages/core/assistant-testing/src/functions/exa/index.ts
@@ -1,0 +1,2 @@
+export { default as exaFunction } from './exa';
+export { default as exaMockFunction } from './mock';
