+++ packages/core/assistant-testing/src/functions/research/index.ts
@@ -2,7 +2,6 @@
 // Copyright 2025 DXOS.org
 //
 
-export * from './exa';
 export { default as research } from './research';
 export * from './graph';
 export * from './research-graph';
