+++ packages/core/assistant-testing/src/functions/research/research.ts
@@ -5,15 +5,26 @@
 import { AiToolkit } from '@effect/ai';
 import { Array, Effect, Layer, Schema } from 'effect';
 
-import { AiService, ConsolePrinter, ToolExecutionService, ToolResolverService } from '@dxos/ai';
-import { AiSession, GenerationObserver } from '@dxos/assistant';
+import { AiService, ConsolePrinter, ToolId } from '@dxos/ai';
+import {
+  AiSession,
+  GenerationObserver,
+  makeToolExecutionServiceFromFunctions,
+  makeToolResolverFromFunctions,
+} from '@dxos/assistant';
 import { Obj } from '@dxos/echo';
-import { ContextQueueService, DatabaseService, TracingService, defineFunction } from '@dxos/functions';
+import {
+  ContextQueueService,
+  DatabaseService,
+  LocalFunctionExecutionService,
+  TracingService,
+  defineFunction,
+} from '@dxos/functions';
 import { type DXN } from '@dxos/keys';
 
-import { ExaToolkit } from './exa';
 import { LocalSearchHandler, LocalSearchToolkit, makeGraphWriterHandler, makeGraphWriterToolkit } from './graph';
 // TODO(dmaretskyi): Vite build bug with instruction files with the same filename getting mixed-up.
+import { exaFunction, exaMockFunction } from '../exa';
 import PROMPT from './instructions-research.tpl?raw';
 import { createResearchGraph, queryResearchGraph } from './research-graph';
 import { ResearchDataTypes } from './types';
@@ -57,13 +68,13 @@ export default defineFunction({
           prompt: query,
           history: [],
           system: PROMPT,
-          toolkit: AiToolkit.merge(ExaToolkit, LocalSearchToolkit, GraphWriterToolkit),
+          toolkit: AiToolkit.merge(LocalSearchToolkit, GraphWriterToolkit),
+          toolIds: [mockSearch ? ToolId.make(exaMockFunction.name) : ToolId.make(exaFunction.name)],
           observer: GenerationObserver.fromPrinter(new ConsolePrinter({ tag: 'research' })),
         })
         .pipe(
           Effect.provide(
             Layer.mergeAll(
-              mockSearch ? ExaToolkit.layerMock : ExaToolkit.layerLive,
               LocalSearchHandler,
               makeGraphWriterHandler(GraphWriterToolkit, { onAppend: (dxns) => newObjectDXNs.push(...dxns) }),
               ContextQueueService.layer(researchQueue),
@@ -85,8 +96,13 @@ export default defineFunction({
     Effect.provide(
       Layer.mergeAll(
         AiService.model('@anthropic/claude-sonnet-4-0'),
-        ToolResolverService.layerEmpty,
-        ToolExecutionService.layerEmpty,
+        makeToolResolverFromFunctions([exaFunction, exaMockFunction], AiToolkit.make()),
+        makeToolExecutionServiceFromFunctions(
+          [exaFunction, exaMockFunction],
+          AiToolkit.make() as any,
+          Layer.empty as any,
+        ),
+        LocalFunctionExecutionService.layer,
       ),
     ),
   ),
